PolyaGammaHybridSamplers

Documentation for PolyaGammaHybridSamplers.

PolyaGammaHybridSamplers.PolyaGammaHybridSampler

PolyaGammaHybridSamplers.rand_pgdevroye

PolyaGammaHybridSamplers.rand_pgnormalapprox

PolyaGammaHybridSamplers.rand_pgsaddlepoint

PolyaGammaHybridSamplers.PolyaGammaHybridSampler — Type

PolyaGammaHybridSampler(b::Integer, z::Real, method::AbstractString)

A sampler for the Polya-Gamma distribution using a hybrid of the saddlepoint approximation, the Devroye method, and the normal approximation, each of which are discussed in Windle et al. (2014) – see README.md for details. The method used is determined by the the method parameter, which defaults to "hybrid" when not specified.

Arguments

b::Integer: The shape parameter of the Polya-Gamma distribution. Must be positive.

z::Real: The exponential tilting parameter of the Polya-Gamma distribution.

method::PGSamplingMethod : An Enum object specifying the method used to sample from the Polya-Gamma distribution. Must be one ofHYBRID,DEVROYE,SADDLEPOINT, orNORMALAPPROX. Defaults toHYBRID` when not specified.

Returns

A PolyaGammaHybridSampler object which can be sampled using rand or rand!.

Examples

julia> using PolyaGammaHybridSamplers
julia> s = PolyaGammaHybridSampler(1, 1.0)
julia> rand(s)

PolyaGammaHybridSamplers.rand_pgdevroye — Method

rand_pgdevroye(b::Integer, z::Real, rng::AbstractRNG)

Sample from a Polya-Gamma distribution using the Devroye method.

Arguments

b::Integer: the shape parameter

z::Real: the exponential tilting parameter

rng::AbstractRNG: random number generator object for rand

Returns

A sample from the Polya-Gamma distribution with shape parameter b and exponential tilting parameter z.

Notes

This method is exact, but is increasingly slower as b increases.

PolyaGammaHybridSamplers.rand_pgnormalapprox — Method

rand_pgnormalapprox(b::Real, z::Real, rng::AbstractRNG)

Sample from a Polya-Gamma distribution using the normal approximation method.

Arguments

b::Real: the shape parameter

z::Real: the exponential tilting parameter

rng::AbstractRNG: random number generator object for rand

Returns

A sample from the Polya-Gamma distribution with shape parameter b and exponential tilting parameter z.

Notes

This method is an approximation, but very efficient for large b.

This method is recommended for b > 170, however no warning is given if b is too small.

This method supports non-integer b.

PolyaGammaHybridSamplers.rand_pgsaddlepoint — Method

rand_pgsaddlepoint(b::Real, z::Real, rng::AbstractRNG)

Sample from a Polya-Gamma distribution using the saddlepoint approximation method.

Arguments

b::Real: the shape parameter

z::Real: the exponential tilting parameter

rng::AbstractRNG: random number generator object for rand

Returns

A sample from the Polya-Gamma distribution with shape parameter b and exponential tilting parameter z.

Notes

This method is an approximation, but very efficient for large b.

This method is recommended for b > 13, however no warning is given if b is too small.

This method supports non-integer b.